requirements.txt
sentence-transformers==2.2.2
huggingface-hub==0.16.4
chromadb==0.4.13
pypdf2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
flask==2.3.2
torch==2.6.0
transformers==4.30.2